+++ changelog.md
@@ -1,5 +1,8 @@
+* 2.4.0
+  - Add Describe and Alter Configs APIs, part of KIP-133
+
 * 2.3.6
-   - Upgrade snappyer and crc32cer to fix build in windows
+  - Upgrade snappyer and crc32cer to fix build in windows
 
 * 2.3.5
   - Improve produce request encoding performance by 35%